+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 
 Latest
 ------
+* Added flag to APILoader so checklists can be updated even if the edited
+  timestamp is unchanged.
 
 0.5.0 (2025-01-20)
 ------------------